fix(datepicker): move code from components/datepicker to index file
this fixes the `require` declarations
